Create landing page, static file structure
This commit contains a new landing that links to the various
applications on the site, as well as the start of an MVC framework
for the server. Database connection and query logic has been moved
to a /model package, and HTML, CSS, and images have been moved to /web.

/web may be renamed to /dist in later iterations, but it was my
understanding that this was not the convention in golang. In order
to increase code extensibility, /web/stylesheets/ contains a common
css reset, as well as theme.css, which includes common fonts, colors,
and classes.